Move contracts from fixtures, migrate to hardhat
Move ethereum-smartcontracts from packages/cli/__fixtures__ to root contracts directory. Migrate deployment scripts from rocketh to hardhat-deploy framework. Update test files from earl to chai testing framework. Standardize code indentation and add Next.js type definitions for base-account-quickstart.
