+++ contracts/.gitattributes
