+++ contracts/.prettierignore
@@ -1,31 +1,17 @@
-############################################
-# .gitignore
-#############################################
-/node_modules
-
-.vscode/*
-!*.default
-*.local
-
-# rocketh
-generated
-deployments/localhost
-deployments/hardhat*
-deployments/lan
-_metadata
-
-# hardhat
-/coverage
-/coverage.json
-/.coverage-config.json
-/.coverage-data.json
-cache
-artifacts
-docs
-
-.vscode/*
-!*.default
-#############################################
-deployments
+export/
+deployments/
+artifacts/
+cache/
+coverage/
+node_modules/
 package.json
+typechain-types/
+dapp_test/test.sol
+_lib/
+.dappp
+*.json
+.yalc
+
+.prettierrc.js
 pnpm-lock.yaml
+package.json
