+++ contracts/deploy/002_deploy_greetings_registry.ts
@@ -1,30 +1,23 @@
-import {deployScript, artifacts} from '../rocketh/deploy.js';
-import {parseEther} from 'viem';
+import {HardhatRuntimeEnvironment} from 'hardhat/types';
+import {DeployFunction} from 'hardhat-deploy/types';
 
-export default deployScript(
-	async (env) => {
-		const {deployer} = env.namedAccounts;
-		const useProxy = !env.tags.live;
+const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
+	const {deployer} = await hre.getNamedAccounts();
+	const {deploy} = hre.deployments;
+	const useProxy = !hre.network.live;
 
-		// proxy only in non-live network (localhost and hardhat network) enabling HCR (Hot Contract Replacement)
-		// in live network, proxy is disabled and constructor is invoked
-		await env.deployViaProxy(
-			'GreetingsRegistry',
-			{
-				account: deployer,
-				artifact: artifacts.GreetingsRegistry,
-				args: ['2'],
-			},
-			{
-				proxyDisabled: !useProxy,
-				execute: 'postUpgrade',
-			},
-		);
+	// proxy only in non-live network (localhost and hardhat network) enabling HCR (Hot Contract Replacement)
+	// in live network, proxy is disabled and constructor is invoked
+	await deploy('GreetingsRegistry', {
+		from: deployer,
+		proxy: useProxy && 'postUpgrade',
+		args: [2],
+		log: true,
+		autoMine: true, // speed up deployment on local network (ganache, hardhat), no effect on live networks
+	});
 
-		return !useProxy; // when live network, record the script as executed to prevent rexecution
-	},
-	{
-		tags: ['GreetingsRegistry'],
-		id: 'deploy_greetings_registry', // id required to prevent reexecution
-	},
-);
+	return !useProxy; // when live network, record the script as executed to prevent rexecution
+};
+export default func;
+func.id = 'deploy_greetings_registry'; // id required to prevent reexecution
+func.tags = ['GreetingsRegistry'];
