+++ contracts/scripts/seed.ts
