+++ contracts/scripts/setMessage.ts
@@ -1,25 +1,17 @@
-import hre from 'hardhat';
-import {Abi_GreetingsRegistry} from '../generated/abis/GreetingsRegistry.js';
-import {loadEnvironmentFromHardhat} from '../rocketh/environment.js';
+import {deployments, getUnnamedAccounts} from 'hardhat';
+const {execute} = deployments;
+// example script
 
 const args = process.argv.slice(2);
 const account = args[0];
 const message = args[1];
 
 async function main() {
-	const env = await loadEnvironmentFromHardhat({hre});
+	const accountAddress = isNaN(parseInt(account)) ? account : (await getUnnamedAccounts())[parseInt(account)];
 
-	const accountAddress = /^0x[a-fA-F0-9]{40}$/.test(account) || !/^\d+$/.test(account)
-		? account
-		: env.unnamedAccounts[parseInt(account)];
-
-	const GreetingsRegistry = env.get<Abi_GreetingsRegistry>('GreetingsRegistry');
-	await env.execute(GreetingsRegistry, {
-		account: accountAddress,
-		functionName: 'setMessage',
-		args: [message || 'hello'],
-	});
+	await execute('GreetingsRegistry', {from: accountAddress, log: true}, 'setMessage', message || 'hello');
 }
+
 main()
 	.then(() => process.exit(0))
 	.catch((error) => {
