+++ contracts/src/ERC20/ERC20Internal.sol
@@ -2,17 +2,13 @@
 pragma solidity 0.8.17;
 
 abstract contract ERC20Internal {
-    function _approveFor(
-        address owner,
-        address target,
-        uint256 amount
-    ) internal virtual;
+	function _approveFor(address owner, address target, uint256 amount) internal virtual;
 
-    function name() public virtual returns (string memory);
+	function name() public virtual returns (string memory);
 
-    function _mint(address to, uint256 amount) internal virtual;
+	function _mint(address to, uint256 amount) internal virtual;
 
-    function _burnFrom(address from, uint256 amount) internal virtual;
+	function _burnFrom(address from, uint256 amount) internal virtual;
 
-    function _internal_totalSupply() internal view virtual returns (uint256);
+	function _internal_totalSupply() internal view virtual returns (uint256);
 }
