+++ contracts/src/ERC20/SimpleERC20.sol
@@ -6,13 +6,13 @@ import "./ERC20Base.sol";
 import "./WithPermitAndFixedDomain.sol";
 
 contract SimpleERC20 is ERC20Base, WithPermitAndFixedDomain {
-    constructor(address to, uint256 amount) WithPermitAndFixedDomain("1") {
-        _mint(to, amount);
-    }
+	constructor(address to, uint256 amount) WithPermitAndFixedDomain("1") {
+		_mint(to, amount);
+	}
 
-    string public constant symbol = "SIMPLE";
+	string public constant symbol = "SIMPLE";
 
-    function name() public pure override returns (string memory) {
-        return "Simple ERC20";
-    }
+	function name() public pure override returns (string memory) {
+		return "Simple ERC20";
+	}
 }
