+++ contracts/src/Interfaces/IERC2612Standalone.sol
@@ -3,17 +3,17 @@
 pragma solidity 0.8.17;
 
 interface IERC2612Standalone {
-    function permit(
-        address owner,
-        address spender,
-        uint256 value,
-        uint256 deadline,
-        uint8 v,
-        bytes32 r,
-        bytes32 s
-    ) external;
+	function permit(
+		address owner,
+		address spender,
+		uint256 value,
+		uint256 deadline,
+		uint8 v,
+		bytes32 r,
+		bytes32 s
+	) external;
 
-    function nonces(address owner) external view returns (uint256);
+	function nonces(address owner) external view returns (uint256);
 
-    function DOMAIN_SEPARATOR() external view returns (bytes32);
+	function DOMAIN_SEPARATOR() external view returns (bytes32);
 }
