+++ contracts/src/Libraries/Constants.sol
@@ -3,7 +3,6 @@
 pragma solidity 0.8.17;
 
 library Constants {
-    uint256 internal constant UINT256_MAX =
-        0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF;
-    uint256 internal constant DECIMALS_18 = 1000000000000000000;
+	uint256 internal constant UINT256_MAX = 0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF;
+	uint256 internal constant DECIMALS_18 = 1000000000000000000;
 }
