+++ contracts/test/GreetingsRegistry.test.ts
@@ -1,49 +1,25 @@
-import {expect} from 'earl';
-import {describe, it} from 'node:test';
-import {network} from 'hardhat';
-import {Abi_GreetingsRegistry} from '../generated/abis/GreetingsRegistry.js';
-import {EthereumProvider} from 'hardhat/types/providers';
-import {loadAndExecuteDeploymentsFromFiles} from '../rocketh/environment.js';
+import {expect} from 'chai';
+import {ethers, deployments, getUnnamedAccounts} from 'hardhat';
+import {GreetingsRegistry} from '../typechain-types';
+import {setupUsers} from './utils';
 
-function setupFixtures(provider: EthereumProvider) {
-	return {
-		async deployAll() {
-			const env = await loadAndExecuteDeploymentsFromFiles({
-				provider: provider,
-			});
-
-			// Deployment are inherently untyped since they can vary from
-			//  network or even be different from current artifacts so here
-			//  we type them manually assuming the artifact is still matching
-			const GreetingsRegistry =
-				env.get<Abi_GreetingsRegistry>('GreetingsRegistry');
-
-			return {
-				env,
-				GreetingsRegistry,
-				namedAccounts: env.namedAccounts,
-				unnamedAccounts: env.unnamedAccounts,
-			};
-		},
+const setup = deployments.createFixture(async () => {
+	await deployments.fixture('GreetingsRegistry');
+	const contracts = {
+		GreetingsRegistry: await ethers.getContract<GreetingsRegistry>('GreetingsRegistry'),
 	};
-}
-
-const {provider, networkHelpers} = await network.connect();
-const {deployAll} = setupFixtures(provider);
-
+	const users = await setupUsers(await getUnnamedAccounts(), contracts);
+	return {
+		...contracts,
+		users,
+	};
+});
 describe('GreetingsRegistry', function () {
 	it('setMessage works', async function () {
-		const {env, GreetingsRegistry, unnamedAccounts} =
-			await networkHelpers.loadFixture(deployAll);
-
+		const {users, GreetingsRegistry} = await setup();
 		const testMessage = 'Hello World';
-
-		await env.execute(GreetingsRegistry, {
-			account: unnamedAccounts[0],
-			functionName: 'setMessage',
-			args: [testMessage],
-		});
-
-		// TODO expect(...).to.emit(GreetingsRegistry, 'MessageChanged').withArgs(users[0].address, testMessage);
+		await expect(users[0].GreetingsRegistry.setMessage(testMessage))
+			.to.emit(GreetingsRegistry, 'MessageChanged')
+			.withArgs(users[0].address, testMessage);
 	});
 });
