+++ contracts/test/SimpleERC20.test.ts
@@ -1,66 +1,34 @@
-import {expect} from 'earl';
-import {describe, it} from 'node:test';
-import {network} from 'hardhat';
-import {EthereumProvider} from 'hardhat/types/providers';
-import {loadAndExecuteDeploymentsFromFiles} from '../rocketh/environment.js';
-import {Abi_SimpleERC20} from '../generated/abis/SimpleERC20.js';
+import {expect} from 'chai';
+import {ethers, deployments, getUnnamedAccounts, getNamedAccounts} from 'hardhat';
+import {IERC20} from '../typechain-types';
+import {setupUser, setupUsers} from './utils';
 
-function setupFixtures(provider: EthereumProvider) {
-	return {
-		async deployAll() {
-			const env = await loadAndExecuteDeploymentsFromFiles({
-				provider: provider,
-			});
-
-			// Deployment are inherently untyped since they can vary from
-			//  network or even be different from current artifacts so here
-			//  we type them manually assuming the artifact is still matching
-			const SimpleERC20 = env.get<Abi_SimpleERC20>('SimpleERC20');
-
-			return {
-				env,
-				SimpleERC20,
-				namedAccounts: env.namedAccounts,
-				unnamedAccounts: env.unnamedAccounts,
-			};
-		},
+const setup = deployments.createFixture(async () => {
+	await deployments.fixture('SimpleERC20');
+	const {simpleERC20Beneficiary} = await getNamedAccounts();
+	const contracts = {
+		SimpleERC20: await ethers.getContract<IERC20>('SimpleERC20'),
 	};
-}
-
-const {provider, networkHelpers} = await network.connect();
-const {deployAll} = setupFixtures(provider);
+	const users = await setupUsers(await getUnnamedAccounts(), contracts);
+	return {
+		...contracts,
+		users,
+		simpleERC20Beneficiary: await setupUser(simpleERC20Beneficiary, contracts),
+	};
+});
 
 describe('SimpleERC20', function () {
 	it('transfer fails', async function () {
-		const {env, SimpleERC20, unnamedAccounts} =
-			await networkHelpers.loadFixture(deployAll);
-
-		await expect(
-			env.execute(SimpleERC20, {
-				account: unnamedAccounts[0],
-				functionName: 'transfer',
-				args: [unnamedAccounts[1], 1n],
-			}),
-		).toBeRejectedWith('NOT_ENOUGH_TOKENS');
+		const {users} = await setup();
+		await expect(users[0].SimpleERC20.transfer(users[1].address, 1)).to.be.revertedWith('NOT_ENOUGH_TOKENS');
 	});
 
 	it('transfer succeed', async function () {
-		const {env, SimpleERC20, unnamedAccounts, namedAccounts} =
-			await networkHelpers.loadFixture(deployAll);
+		const {users, simpleERC20Beneficiary, SimpleERC20} = await setup();
+		await simpleERC20Beneficiary.SimpleERC20.transfer(users[1].address, 1);
 
-		await env.execute(SimpleERC20, {
-			account: namedAccounts.simpleERC20Beneficiary,
-			functionName: 'transfer',
-			args: [unnamedAccounts[1], 1n],
-		});
-
-		env.execute(SimpleERC20, {
-			account: namedAccounts.simpleERC20Beneficiary,
-			functionName: 'transfer',
-			args: [unnamedAccounts[1], 1n],
-		});
-		// TODO
-		// expect(...).toEmit(SimpleERC20, 'Transfer')
-		// .withArgs(simpleERC20Beneficiary.address, users[1].address, 1));
+		await expect(simpleERC20Beneficiary.SimpleERC20.transfer(users[1].address, 1))
+			.to.emit(SimpleERC20, 'Transfer')
+			.withArgs(simpleERC20Beneficiary.address, users[1].address, 1);
 	});
 });
