+++ contracts/tsconfig.json
@@ -1,17 +1,18 @@
 {
-	"compilerOptions": {
-		"lib": ["es2023"],
-		"module": "node16",
-		"target": "es2022",
-		"moduleResolution": "node16",
-		"strict": true,
-		"esModuleInterop": true,
-		"skipLibCheck": true,
-		"sourceMap": true,
-		"declaration": true,
-		"declarationMap": true,
-		"outDir": "dist",
-		"rootDir": "."
-	},
-	"include": ["deploy", "generated"]
+  "compilerOptions": {
+    "target": "es5",
+    "module": "commonjs",
+    "strict": true,
+    "esModuleInterop": true,
+    "moduleResolution": "node",
+    "forceConsistentCasingInFileNames": true,
+    "outDir": "dist"
+  },
+  "include": [
+    "hardhat.config.ts",
+    "./scripts",
+    "./deploy",
+    "./test",
+    "typechain/**/*"
+  ]
 }
