--- packages/cli/src/plugins/__fixtures__/ethereum-smartcontracts/.gitignore
@@ -1,23 +1,0 @@
-cache/
-artifacts/
-
-coverage*
-typechain-types/
-
-.vscode/*
-!.vscode/settings.json.default
-!.vscode/launch.json.default
-!.vscode/extensions.json.default
-
-node_modules/
-.env
-
-.prettierrc.js
-pnpm-lock.yaml
-package.json
-
-contractsInfo.json
-deployments/hardhat
-deployments/localhost
-
-/.dapp
