--- packages/cli/src/plugins/__fixtures__/ethereum-smartcontracts/.prettierignore
@@ -1,17 +1,0 @@
-export/
-deployments/
-artifacts/
-cache/
-coverage/
-node_modules/
-package.json
-typechain-types/
-dapp_test/test.sol
-_lib/
-.dappp
-*.json
-.yalc
-
-.prettierrc.js
-pnpm-lock.yaml
-package.json
