--- packages/cli/src/plugins/__fixtures__/ethereum-smartcontracts/deploy/001_deploy_simple_erc20.ts
@@ -1,19 +1,0 @@
-import {HardhatRuntimeEnvironment} from 'hardhat/types';
-import {DeployFunction} from 'hardhat-deploy/types';
-import {parseEther} from 'ethers';
-
-const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
-	const {deployments, getNamedAccounts} = hre;
-	const {deploy} = deployments;
-
-	const {deployer, simpleERC20Beneficiary} = await getNamedAccounts();
-
-	await deploy('SimpleERC20', {
-		from: deployer,
-		args: [simpleERC20Beneficiary, parseEther('1000000000')],
-		log: true,
-		autoMine: true, // speed up deployment on local network (ganache, hardhat), no effect on live networks
-	});
-};
-export default func;
-func.tags = ['SimpleERC20'];
