--- packages/cli/src/plugins/__fixtures__/ethereum-smartcontracts/src/Interfaces/IERC2612Standalone.sol
@@ -1,19 +1,0 @@
-// SPDX-License-Identifier: AGPL-1.0
-
-pragma solidity 0.8.17;
-
-interface IERC2612Standalone {
-	function permit(
-		address owner,
-		address spender,
-		uint256 value,
-		uint256 deadline,
-		uint8 v,
-		bytes32 r,
-		bytes32 s
-	) external;
-
-	function nonces(address owner) external view returns (uint256);
-
-	function DOMAIN_SEPARATOR() external view returns (bytes32);
-}
