--- packages/cli/src/plugins/__fixtures__/ethereum-smartcontracts/src/Libraries/Constants.sol
@@ -1,8 +1,0 @@
-// SPDX-License-Identifier: AGPL-1.0
-
-pragma solidity 0.8.17;
-
-library Constants {
-	uint256 internal constant UINT256_MAX = 0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF;
-	uint256 internal constant DECIMALS_18 = 1000000000000000000;
-}
